fix more build errors
