+++ frontend/src/types.ts
@@ -21,20 +21,6 @@ export interface updateBookingHoursBody {
   endWeek?: number;
 }
 
-export interface ProjectWithCustomerModel {
-  projectName: string;
-  customerName: string;
-  projectState: ProjectState;
-  isBillable: boolean;
-}
-
-export interface EngagementWriteModel {
-  projectState?: ProjectState;
-  isBillable?: boolean;
-  projectName?: string;
-  customerName?: string;
-}
-
 export interface updateProjectStateBody {
   engagementId: string;
   projectState: ProjectState;
